+++ examples/convert_ieeg_to_bids.py
@@ -7,8 +7,6 @@ In this example, we use MNE-BIDS to create a BIDS-compatible directory of iEEG
 data. Specifically, we will follow these steps:
 
 1. Download some iEEG data from the
-   `PhysioBank database <https://physionet.org/physiobank/database>`_
-   and the
    `MNE-ECoG ex <https://mne.tools/stable/auto_tutorials/misc/plot_ecog>`_.
 
 2. Load the data, extract information, and save in a new BIDS directory.
@@ -37,15 +35,14 @@ refer to the iEEG-BIDS specification.
 # License: BSD (3-clause)
 
 import os
-import tempfile
 from pprint import pprint
 
 import numpy as np
-from scipy.io import loadmat
 
 import mne
 from mne_bids import write_raw_bids, make_bids_basename, read_raw_bids
 from mne_bids.utils import print_dir_tree
+from mne_bids.tsv_handler import _from_tsv
 
 ###############################################################################
 # Step 1: Download the data
@@ -59,33 +56,39 @@ from mne_bids.utils import print_dir_tree
 # Conveniently, there is already a data loading function available with
 # MNE-Python:
 
-# The electrode coords data are in the Matlab format: '.mat'.
-# This is easy to read in with :func:`scipy.io.loadmat` function.
-mat = loadmat(mne.datasets.misc.data_path() + '/ecog/sample_ecog.mat')
-ch_names = mat['ch_names'].tolist()
-ch_names = [x.strip() for x in ch_names]
-elec = mat['elec']  # electrode positions given in meters
+misc_path = mne.datasets.misc.data_path()
+
+# The electrode coords data are in the tsv file format
+electrode_tsv = _from_tsv(misc_path + '/ecog/sample_ecog_electrodes.tsv')
+ch_names = electrode_tsv['name']
+
+# load in the xyz coordinates as a float
+elec = np.empty(shape=(len(ch_names), 3))
+for ind, axis in enumerate(['x', 'y', 'z']):
+    elec[:, ind] = list(map(float, electrode_tsv[axis]))
 
 ###############################################################################
 # Now we make a montage stating that the iEEG contacts are in MRI
 # coordinate system.
 montage = mne.channels.make_dig_montage(ch_pos=dict(zip(ch_names, elec)),
-                                        coord_frame='head')
+                                        coord_frame='mri')
 print('Created %s channel positions' % len(ch_names))
 print(dict(zip(ch_names, elec)))
 
 ###############################################################################
-# We also need to get some sample iEEG data,
-# so we will just generate random data from white noise.
-# Here is where you would use your own data if you had it.
-ieegdata = np.random.rand(len(ch_names), 1000)
-
-# We will create a :func:`mne.io.RawArray` object and
+# We will load a :func:`mne.io.Raw` object and
 # use the montage we created.
 info = mne.create_info(ch_names, 1000., 'ecog')
-raw = mne.io.RawArray(ieegdata, info)
+raw = mne.io.read_raw_edf(misc_path + '/ecog/sample_ecog.edf')
 raw.set_channel_types({ch: 'ecog' for ch in raw.ch_names})
-raw.set_montage(montage)
+
+# set the bad channels
+raw.info['bads'].extend(['BTM1', 'BTM2', 'BTM3', 'BTM4', 'BTM5', 'BTM6',
+                         'BTP1', 'BTP2', 'BTP3', 'BTP4', 'BTP5', 'BTP6',
+                         'EKGL', 'EKGR'])
+
+# set montage
+raw.set_montage(montage, on_missing='warn')
 
 ###############################################################################
 # Let us confirm what our channel coordinates look like.
@@ -147,15 +150,9 @@ bids_basename = make_bids_basename(subject=subject_id,
                                    task=task,
                                    acquisition="ecog")
 
-# need to set the filenames if we are initializing data from array
-with tempfile.TemporaryDirectory() as tmp_root:
-    tmp_fpath = os.path.join(tmp_root, "test_raw.fif")
-    raw.save(tmp_fpath)
-    raw = mne.io.read_raw_fif(tmp_fpath)
-
-    # write `raw` to BIDS and anonymize it into BrainVision format
-    write_raw_bids(raw, bids_basename, bids_root=bids_root,
-                   anonymize=dict(daysback=30000), overwrite=True)
+# write `raw` to BIDS and anonymize it into BrainVision format
+write_raw_bids(raw, bids_basename, bids_root=bids_root,
+               anonymize=dict(daysback=30000), overwrite=True)
 
 ###############################################################################
 # Step 3: Check and compare with standard
@@ -202,6 +199,9 @@ dig = [raw.info['dig'][pick] for pick in picks]
 chs = [raw.info['chs'][pick] for pick in picks]
 pos = np.array([ch['r'] for ch in dig[:5]])
 ch_names = np.array([ch['ch_name'] for ch in chs[:5]])
+
+print("The channel montage after writing into BIDS: ")
+pprint(dig[0:5])
 print("The channel coordinates after writing into BIDS: ")
 pprint([x for x in zip(ch_names, pos)])